+++ gather/config_test.go
@@ -24,6 +24,7 @@ func TestLoadingConfig(t *testing.T) {
 	assert.Equal(t, cfg.GetString(gather.CK_server_datastore_ip.String()), "http://localhost")
 	assert.Equal(t, cfg.GetBool(gather.CK_args_debug.String()), true)
 	assert.Equal(t, cfg.GetString(gather.CK_args_admindb_name.String()), "adh-admin")
+	assert.Equal(t, cfg.GetInt(gather.CK_server_monitoring_port.String()), 9191)
 }
 
 func TestLoadingDefaults(t *testing.T) {
@@ -38,6 +39,7 @@ func TestLoadingDefaults(t *testing.T) {
 	assert.Equal(t, cfg.GetString(gather.CK_server_datastore_ip.String()), "http://localhost")
 	assert.Equal(t, cfg.GetBool(gather.CK_args_debug.String()), false)
 	assert.Equal(t, cfg.GetString(gather.CK_args_admindb_name.String()), "adh-admin")
+	assert.Equal(t, cfg.GetInt(gather.CK_server_monitoring_port.String()), 9191)
 }
 
 func TestWithEnvironmentVariables(t *testing.T) {